Screenshot2
When clicked on restaurant type dessert parlor it gave me top 10 restaurants which are having this type. It gave me other details as well like do i have an option of booking a table. Unfortunately, we don't have option to book a table for this case. And avg rating for dessert parlor is 3.88 . It gave me location and cuisines as well . So this report pretty well answers our questions. According to the business requirement we can build other dashboard as well.